refactor trtllm attention quant fusion
