+++ tensorrt_llm/_torch/auto_deploy/custom_ops/attention_interface.py
@@ -31,7 +31,7 @@ from typing import Dict, List, Literal, Optional, Protocol, Sequence, Set, Tuple
 import numpy as np
 import torch
 from torch._ops import OpOverloadPacket
-from torch.fx import Node
+from torch.fx import GraphModule, Node
 from torch.types import Number
 
 from tensorrt_llm.llmapi.llm_args import KvCacheConfig
@@ -40,22 +40,6 @@ from ...._utils import nvtx_range, prefer_pinned, str_dtype_to_torch
 from ..utils.logger import ad_logger
 
 Constant = Union[int, float, str, None]
-
-_TRTLLM_ATTN_FP8_INPUT_SCALE_KEY = "trtllm_attention_input_scale"
-
-
-def set_trtllm_attention_fp8_input_scale(attn_node: Node, input_scale: Node) -> None:
-    attn_node.meta[_TRTLLM_ATTN_FP8_INPUT_SCALE_KEY] = input_scale
-
-
-def get_trtllm_attention_fp8_input_scale(attn_node: Node) -> Optional[Node]:
-    scale = attn_node.meta.get(_TRTLLM_ATTN_FP8_INPUT_SCALE_KEY)
-    return scale if isinstance(scale, Node) else None
-
-
-def clear_trtllm_attention_fp8_input_scale(attn_node: Node) -> None:
-    attn_node.meta.pop(_TRTLLM_ATTN_FP8_INPUT_SCALE_KEY, None)
-
 
 # Torch dtype → numpy dtype for fast list-to-tensor conversion.
 # numpy's list→array conversion is ~2-3x faster than torch.tensor(list) for large lists.
@@ -1582,6 +1566,15 @@ class AttentionDescriptor(ABC):
         This method is responsible for preparing the attention op for the forward pass.
         This function is not expected to be graph capturable or compatible with cuda graphs. It can
         use any argument from the SequenceInfo interface as input argument to its function.
+        """
+        return None
+
+    @classmethod
+    def prepare_node_for_cache_insertion(cls, gm: GraphModule, attn_node: Node) -> None:
+        """Perform optional backend-specific graph prep before cached attention insertion.
+
+        Default implementation is a no-op. Backends can override this hook to materialize
+        backend-specific graph nodes or metadata needed by ``get_constants``.
         """
         return None
